Объединил несколько активити в одно - MainActivity
