+++ app/src/main/java/com/example/university/View/Main/MainScreens.kt
@@ -1,0 +1,7 @@
+package com.example.university.View.Main
+
+sealed class MainScreens(val route: String) {
+    object Main : MainScreens("main_screen")
+    object AddNew : MainScreens("add_screen")
+    object Settings : MainScreens("settings_screen")
+}
