+++ app/src/main/java/com/example/university/ViewModel/DI/dataModule.kt
@@ -1,9 +1,9 @@
 package com.example.university.ViewModel.DI
 
-import android.content.SharedPreferences
-import android.preference.PreferenceManager
 import com.example.university.Model.DBManager
 import com.example.university.Model.MySharedPreferences
+import com.example.university.ViewModel.MainActivityViewModel
+import com.example.university.ViewModel.AddViewModel
 import com.example.university.ViewModel.MainViewModel
 import com.example.university.ViewModel.SettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -19,15 +19,18 @@ private fun Module.data(){
     single<DBManager>{
         DBManager(context = get())
     }
-    /*single<SharedPreferences>{
-        PreferenceManager.getDefaultSharedPreferences(get())
-    }*/
     single<MySharedPreferences>{
         MySharedPreferences(context = get())
     }
 }
 
 private fun Module.presentation(){
+    viewModel<MainActivityViewModel> {
+        MainActivityViewModel(
+            db = get(),
+            msp = get()
+        )
+    }
     viewModel<SettingsViewModel> {
         SettingsViewModel(
             db = get(),
@@ -36,6 +39,12 @@ private fun Module.presentation(){
     }
     viewModel<MainViewModel> {
         MainViewModel(
+            db = get(),
+            msp = get()
+        )
+    }
+    viewModel<AddViewModel>{
+        AddViewModel(
             db = get(),
             msp = get()
         )
